requirements.txt
langchain-google-genai
langchain-chroma
langchain
langchain-community
python-dotenv
Django